requirements.txt
tabulate==0.7.5
pandas==0.18.1
openpyxl==2.4.0
python-slugify==1.2.1
